+++ Products/CMFPlone/PloneTool.py
@@ -60,7 +60,7 @@ __show__.on()
 
 
 AllowSendto = 'Allow sendto'
-permissions.setDefaultRoles(AllowSendto, ('Anonymous', 'Manager', ))
+permissions.setDefaultRoles(AllowSendto, ('Authenticated', 'Manager', ))
 
 _marker = utils._marker
 _icons = {}